Delete http_requests.py module
The module has only one function that could be placed to general.py
module where it fits more with other requests related stuff.
